requirements.txt
Flask==3.0.2
fpdf2==2.7.8
num2words==0.3.0
gunicorn==21.2.0
